% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{get_lists}
\alias{get_lists}
\title{Fetch all the lists on a SharePoint site}
\usage{
get_lists(site_id, token)
}
\arguments{
\item{site_id}{Site ID}

\item{token}{httr auth token}
}
\value{
A somewhat messy list structure :(
}
\description{
Fetch all the lists on a SharePoint site
}
\examples{
NULL
}
