% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{get_sites}
\alias{get_sites}
\title{Fetch a tibble of SharePoint sites}
\usage{
get_sites(searchquery = NULL, token = msgraph_read_token())
}
\arguments{
\item{searchquery}{Optional search}

\item{token}{httr oauth token}
}
\value{
A tibble
}
\description{
Fetch a tibble of SharePoint sites
}
\examples{
NULL
}
