% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{msgraph_api}
\alias{msgraph_api}
\title{Execute a GET request against the MS Graph API}
\usage{
msgraph_api(
  path,
  msgraph_token,
  query = NULL,
  api_version = default_api_version()
)
}
\arguments{
\item{path}{API endpoint to retrieve}

\item{msgraph_token}{Oauth2 token as returned from `httr::oauth2_token()`}

\item{query}{Named list of parameters to use as the URL query}

\item{api_version}{Version of the Graph API to call}
}
\value{
A msgraph_api object
}
\description{
Calls bare API end points and returns a wrapped `httr` response object
to a higher level function for presentation.
}
\examples{
NULL
}
