% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{msgraph_auth}
\alias{msgraph_auth}
\title{Perform the OAuth 2.0 dance with MS Graph}
\usage{
msgraph_auth(client_id, client_secret)
}
\arguments{
\item{client_id}{Application Id from MS Graph}

\item{client_secret}{Application password from MS Graph}
}
\value{
A `httr` oauth token object
}
\description{
Perform the OAuth 2.0 dance with MS Graph
}
\examples{
NULL
}
