% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{msgraph_token}
\alias{msgraph_token}
\title{Find the location of the OAuth token cache}
\usage{
msgraph_token()
}
\value{
Path to the on-disk OAuth cache
}
\description{
Find the location of the OAuth token cache
}
